refactor: extract ValidationError and controller helpers to shared modules
