+++ src/lib/errors.ts
@@ -1,0 +1,6 @@
+export class ValidationError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'ValidationError';
+  }
+}
